+++ apps/client/src/components/ui/sonner.tsx
@@ -1,11 +1,14 @@
 import type { ToasterProps } from 'sonner'
 import { CircleCheckIcon, InfoIcon, Loader2Icon, OctagonXIcon, TriangleAlertIcon } from 'lucide-react'
+import { useTheme } from 'next-themes'
 import { Toaster as Sonner } from 'sonner'
 
 function Toaster({ ...props }: ToasterProps) {
+  const { resolvedTheme } = useTheme()
+
   return (
     <Sonner
-      theme="light"
+      theme={resolvedTheme === 'dark' ? 'dark' : 'light'}
       className="toaster group"
       icons={{
         success: (